refactors scoring conversations
